features/changelog : historical journal consumption.
Facilitates Glusterfs with the ability to detect file-operations
happened in past by scanning the back-end(brick-level) glusterfs
journal (changelog).

Design:
  * List of changelogs produces in one perfectly running session are
    stored in htime file which also holds necessary information about
    the session start and end time.
  * Involves fixed sized seeks to identify N'th changelog in the list.
  * Requires O(log n), (where n is number of changelogs in the list),
    time to identify the end changelog for the given start-end time
    interval.

Currently the background processing of changelogs is sub optimal. BZ
1097041 tracks the development effort.

For complete design, refer the below link:
http://lists.nongnu.org/archive/html/gluster-devel/2014-02/msg00206.html

BUG: 1091961
Reviewed-on: http://review.gluster.org/6930
Tested-by: Gluster Build System <jenkins@build.gluster.com>

+++ xlators/features/changelog/lib/src/gf-changelog-process.c
@@ -459,8 +459,49 @@ gf_changelog_decode (xlator_t *this, gf_changelog_t *gfc, int from_fd,
         return ret;
 }
 
-static int
-gf_changelog_consume (xlator_t *this, gf_changelog_t *gfc, char *from_path)
+int
+gf_changelog_publish (xlator_t *this, gf_changelog_t *gfc, char *from_path)
+{
+        int         ret        = 0;
+        char dest[PATH_MAX]    = {0,};
+        char to_path[PATH_MAX] = {0,};
+        struct stat stbuf      = {0,};
+
+        (void) snprintf (to_path, PATH_MAX, "%s%s",
+                         gfc->gfc_current_dir, basename (from_path));
+
+        /* handle zerob file that wont exist in current */
+        ret = stat (from_path, &stbuf);
+        if (ret)
+                goto out;
+
+        if (stbuf.st_size == 0) {
+                ret = unlink (from_path);
+                if (ret)
+                        gf_log (this->name, GF_LOG_ERROR,
+                                "could not unlink %s (reason %s)",
+                                from_path, strerror (errno));
+                goto out;
+        }
+
+        (void) snprintf (dest, PATH_MAX, "%s%s",
+                         gfc->gfc_processing_dir, basename (from_path));
+
+        ret = rename (to_path, dest);
+        if (ret){
+                gf_log (this->name, GF_LOG_ERROR,
+                        "error moving %s to processing dir"
+                        " (reason: %s)", to_path, strerror (errno));
+        }
+
+out:
+        return ret;
+}
+
+int
+gf_changelog_consume (xlator_t *this,
+                      gf_changelog_t *gfc,
+                      char *from_path, gf_boolean_t no_publish)
 {
         int         ret        = -1;
         int         fd1        = 0;
@@ -472,6 +513,7 @@ gf_changelog_consume (xlator_t *this, gf_changelog_t *gfc, char *from_path)
 
         ret = stat (from_path, &stbuf);
         if (ret || !S_ISREG(stbuf.st_mode)) {
+                ret = -1;
                 gf_log (this->name, GF_LOG_ERROR,
                         "stat failed on changelog file: %s", from_path);
                 goto out;
@@ -506,6 +548,8 @@ gf_changelog_consume (xlator_t *this, gf_changelog_t *gfc, char *from_path)
                 if (!ret) {
                         /* move it to processing on a successfull
                            decode */
+                        if (no_publish == _gf_true)
+                                goto close_fd;
                         ret = rename (to_path, dest);
                         if (ret)
                                 gf_log (this->name, GF_LOG_ERROR,
@@ -516,10 +560,15 @@ gf_changelog_consume (xlator_t *this, gf_changelog_t *gfc, char *from_path)
 
                 /* remove it from .current if it's an empty file */
                 if (zerob) {
+                        if (no_publish == _gf_true) {
+                                ret = 0;
+                                goto close_fd;
+                        }
+
                         ret = unlink (to_path);
                         if (ret)
                                 gf_log (this->name, GF_LOG_ERROR,
-                                        "could not unlink %s (reason: %s",
+                                        "could not unlink %s (reason: %s)",
                                         to_path, strerror (errno));
                 }
         }
@@ -546,7 +595,7 @@ gf_changelog_ext_change (xlator_t *this,
                         alo = 1;
                         gf_log (this->name, GF_LOG_DEBUG,
                                 "processing changelog: %s", path);
-                        ret = gf_changelog_consume (this, gfc, path);
+                        ret = gf_changelog_consume (this, gfc, path, _gf_false);
                 }
 
                 if (ret)
